Remove player, board, cell class from main.rb file
